+++ projects/ui-components/lf-user-feedback/feedback-image-upload/feedback-image-upload.component.ts
@@ -192,8 +192,8 @@ export class FeedbackImageUploadComponent {
 
 /** @internal */
 enum ImageUploadErrorType {
-  'TooLarge',
-  'UnsupportedFormat',
+  TooLarge,
+  UnsupportedFormat,
 }
 
 /** @internal */